rename schame.ts to schemas.ts
