+++ packages/generator/templates/mutations/update__ModelName__.ts
@@ -1,6 +1,6 @@
 import { resolver } from "@blitzjs/rpc"
 import db from "__prismaFolder__"
-import { Update__ModelName__Schema } from "../schema"
+import { Update__ModelName__Schema } from "../schemas"
 
 export default resolver.pipe(
   resolver.zod(Update__ModelName__Schema),
