+++ packages/generator/templates/validations/schemas.ts
